add eslint

+++ client/src/app/core/services/user.service.ts
@@ -9,7 +9,9 @@ import { BehaviorSubject, Observable, Subject, distinctUntilChanged } from 'rxjs
 })
 export class UserService {
 	public readonly userSubject: Subject<User | false> = new BehaviorSubject<User | false>(false);
-	public readonly user: Observable<User | false> = this.userSubject.asObservable().pipe(distinctUntilChanged());
+	public readonly user: Observable<User | false> = this.userSubject
+		.asObservable()
+		.pipe(distinctUntilChanged());
 
 	public constructor(
 		private readonly api: ApiService,
@@ -19,10 +21,9 @@ export class UserService {
 	public fetchInfo(): void {
 		const authorization = this.accessTokenService.getAccessToken();
 
-		if (!authorization)
-			return this.destroyAuth();
+		if (!authorization) return this.destroyAuth();
 
-		this.api.get<User>(`/api/user`, { headers: { authorization } }).subscribe({
+		this.api.get<User>('/api/user', { headers: { authorization } }).subscribe({
 			next: this.setAuth,
 			error: this.destroyAuth,
 		});
@@ -41,7 +42,7 @@ export class UserService {
 
 	public logIn(code: string): void {
 		this.api.post<{ code: string }>('/auth', { code }).subscribe(({ code: accessToken }) => {
-			console.log('got token', accessToken)
+			console.log('got token', accessToken);
 			this.accessTokenService.saveAccessToken(accessToken);
 			this.fetchInfo();
 		});
